+++ jest.config.js
@@ -1,0 +1,21 @@
+/** @type {import('ts-jest').JestConfigWithTsJest} */
+module.exports = {
+    preset: 'ts-jest',
+    testEnvironment: 'node',
+    roots: ['<rootDir>/src/test'],
+    testMatch: ['**/*.test.ts'],
+    moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
+    collectCoverageFrom: [
+        'src/**/*.ts',
+        '!src/test/**/*.ts'
+    ],
+    coverageDirectory: 'coverage',
+    transform: {
+        '^.+\\.tsx?$': ['ts-jest', {
+            tsconfig: 'tsconfig.test.json'
+        }]
+    },
+    moduleNameMapper: {
+        '^vscode$': '<rootDir>/src/test/__mocks__/vscode.ts'
+    }
+};